refactor: mainから一部をcontrollerに切り出し
